Added content

+++ CPP/GameDev/FirstGame/FirstGame/Engine/Engine.cpp
@@ -7,9 +7,9 @@
 int Engine::SCREEN_WIDTH = 1024;
 int Engine::SCREEN_HEIGHT = 768;
 GLFWwindow* Engine::window = NULL;
-double Engine::dt = 0;
+float Engine::dt = 0;
 
-double Engine::GetDT() {
+float Engine::GetDT() {
 	return dt;
 }
 
@@ -66,7 +66,7 @@ bool Engine::Initialize(const char* windowTitle) {
 	glEnable(GL_BLEND);
 	glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
 
-	lastTime = glfwGetTime();
+	lastTime = (float)glfwGetTime();
 	
 	return true;
 }
@@ -80,7 +80,7 @@ void Engine::UseJoysticks() {
 }
 
 void Engine::Update() {
-	double now = glfwGetTime();
+	float now = (float)glfwGetTime();
 	dt = (now - lastTime);
 	lastTime = now;
